requirements.txt
Flask
python-dotenv
PyYAML
Flask-Cors
gunicorn
